Add a gulp task for TS compilation.
This task compiles the TS files into JS files and places them
in the output folder. For now, it's called dist for distribution but
we can change it if necessary.

The task is configured to use the tsconfig files so we get the compiler
errors. But sourcemaps have to be generated explicitly. The gulp plugin
for typescript doesn't support sourcemaps directly. But there is another
plugin that does so we just use that.

Refs: #1