Add a task for copying html files.
This task will copy the html files from the source folder into the output
folder so the app can be run.

We treat the html files separately from the TS since we may want to do some
preprocessing to them in the future. Also, this task can later be expanded to
copy over other resources such as images and CSS.

REFS: #1