Add a gulp task for linting.
Linting is a kind of static checking of code. It detects all sorts of potential
errors. Very useful for weakly runtime languages like JS and friends.

The linter we are running is configured using a rules file (tslint.json) that
defines what to check for. We are starting with a set of core rules enabled but
no custom rules of our own.

For more about the rules see: https://github.com/palantir/tslint#core-rules

REFS #1