Re-enable top bar.
The topbar is not for tab navigation since all navigation needs to be
fired from the proj browser since the main views need a loaded project element.

So, the topbar will simply inform about the currently selected view. For ease
of use, we also introduce a function to set this title.

REFS: #